+++ product/server/codekvast-server/src/main/java/se/crisp/codekvast/server/codekvast_server/messaging/WebSocketDataHandler.java
@@ -12,18 +12,13 @@ import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 import se.crisp.codekvast.server.codekvast_server.exception.CodekvastException;
 import se.crisp.codekvast.server.codekvast_server.model.event.display.WebSocketMessage;
-import se.crisp.codekvast.server.codekvast_server.model.event.rest.GetMethodUsageRequest;
-import se.crisp.codekvast.server.codekvast_server.model.event.rest.GetMethodUsageResponse;
-import se.crisp.codekvast.server.codekvast_server.model.event.rest.MethodUsageEntry;
-import se.crisp.codekvast.server.codekvast_server.model.event.rest.OrganisationSettings;
+import se.crisp.codekvast.server.codekvast_server.model.event.rest.*;
 import se.crisp.codekvast.server.codekvast_server.service.UserService;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
 import java.security.Principal;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
+import java.util.*;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -74,12 +69,10 @@ public class WebSocketDataHandler extends AbstractEventBusSubscriber {
      */
     @Subscribe
     public void onWebSocketMessage(WebSocketMessage message) {
-        for (String username : message.getUsernames()) {
-            if (webSocketUserPresenceHandler.isPresent(username)) {
-                log.debug("Sending {} to '{}'", message, username);
-                messagingTemplate.convertAndSendToUser(username, "/queue/data", message);
-            }
-        }
+        message.getUsernames().stream().filter(username -> webSocketUserPresenceHandler.isPresent(username)).forEach(username -> {
+            log.debug("Sending {} to '{}'", message, username);
+            messagingTemplate.convertAndSendToUser(username, "/queue/data", message);
+        });
     }
 
     /**
@@ -117,13 +110,41 @@ public class WebSocketDataHandler extends AbstractEventBusSubscriber {
         long now = System.currentTimeMillis();
 
         List<MethodUsageEntry> methods = new ArrayList<>();
-        for (int i = 0; i < request.getPreviewRows(); i++) {
-            long usedAtMillis = random.nextBoolean() ? now - i * 60_000L : 0L;
-            methods.add(MethodUsageEntry.builder().name(String.format("%s-%04d", randomString(80), i)).invokedAtMillis(usedAtMillis).build
-                    ());
+        for (int i = 0; i < request.getMaxPreviewRows(); i++) {
+            boolean dead = request.getMethodUsageScopes().contains(MethodUsageScope.DEAD) ? random.nextBoolean() : false;
+            long usedAtMillis = dead ? 0L : now - i * 60_000L;
+            methods.add(MethodUsageEntry.builder()
+                                        .name(String.format("%s-%04d", randomString(80), i))
+                                        .scope(getRandomScope(request.getMethodUsageScopes(), usedAtMillis))
+                                        .invokedAtMillis(usedAtMillis)
+                                        .build());
         }
 
-        return GetMethodUsageResponse.builder().methods(methods).build();
+        Map<MethodUsageScope, Integer> methodsByScope = new HashMap<>();
+        for (MethodUsageScope scope : request.getMethodUsageScopes()) {
+            methodsByScope.put(scope, random.nextInt(methods.size()));
+        }
+
+        return GetMethodUsageResponse.builder()
+                                     .request(request)
+                                     .methods(methods)
+                                     .numMethods((int) (methods.size() * (1.5d + random.nextDouble())))
+                                     .numMethodsByScope(methodsByScope)
+                                     .build();
+    }
+
+    private MethodUsageScope getRandomScope(Collection<MethodUsageScope> requestScopes, long usedAtMillis) {
+        if (usedAtMillis == 0L && requestScopes.contains(MethodUsageScope.DEAD)) {
+            return MethodUsageScope.DEAD;
+        }
+
+        MethodUsageScope[] values = MethodUsageScope.values();
+        while (true) {
+            MethodUsageScope scope = values[random.nextInt(values.length)];
+            if (scope != MethodUsageScope.DEAD && requestScopes.contains(scope)) {
+                return scope;
+            }
+        }
     }
 
     private final Random random = new Random();