+++ product/server/codekvast-server/src/main/java/se/crisp/codekvast/server/codekvast_server/model/event/rest/GetMethodUsageResponse.java
@@ -1,9 +1,11 @@
 package se.crisp.codekvast.server.codekvast_server.model.event.rest;
 
 import lombok.Builder;
+import lombok.NonNull;
 import lombok.Value;
 
 import java.util.Collection;
+import java.util.Map;
 
 /**
  * @author olle.hallin@crisp.se
@@ -11,14 +13,11 @@ import java.util.Collection;
 @Value
 @Builder
 public class GetMethodUsageResponse {
-    int allMethods;
-    int matchedMethods;
+    @NonNull
+    private final GetMethodUsageRequest request;
 
-    int deadMethods;
-    int probablyDeadInvoked;
-    int bootMethods;
-    int liveMethods;
-    int rowLimit;
+    private final int numMethods;
+    private final Map<MethodUsageScope, Integer> numMethodsByScope;
 
-    Collection<MethodUsageEntry> methods;
+    private final Collection<MethodUsageEntry> methods;
 }