+++ product/server/codekvast-server/src/main/java/se/crisp/codekvast/server/codekvast_server/model/event/rest/MethodUsageEntry.java
@@ -10,5 +10,6 @@ import lombok.Value;
 @Builder
 public class MethodUsageEntry {
     String name;
+    MethodUsageScope scope;
     long invokedAtMillis;
 }