+++ product/server/codekvast-server/src/main/java/se/crisp/codekvast/server/codekvast_server/model/event/rest/MethodUsageScope.java
@@ -1,0 +1,8 @@
+package se.crisp.codekvast.server.codekvast_server.model.event.rest;
+
+/**
+ * @author olle.hallin@crisp.se
+ */
+public enum MethodUsageScope {
+    DEAD, PROBABLY_DEAD, BOOTSTRAP, LIVE;
+}